@@ -2,10 +2,11 @@ package mongoDummy
 
 import (
 	"context"
+
 	mongoModel "github.com/conacry/go-platform/pkg/mongo/model"
 
-	"go.mongodb.org/mongo-driver/mongo"
-	"go.mongodb.org/mongo-driver/mongo/options"
+	"go.mongodb.org/mongo-driver/v2/mongo"
+	"go.mongodb.org/mongo-driver/v2/mongo/options"
 )
 
 type RepositoryDummy struct{}
@@ -14,72 +15,136 @@ func GetRepository() *RepositoryDummy {
 	return &RepositoryDummy{}
 }
 
-func (m *RepositoryDummy) Insert(ctx context.Context, collection mongoModel.Collection, data interface{}) (string, error) {
+func (m *RepositoryDummy) Insert(
+	ctx context.Context,
+	collection mongoModel.Collection,
+	data any,
+) (string, error) {
 	return "", nil
 }
 
-func (m *RepositoryDummy) InsertMany(ctx context.Context, collection mongoModel.Collection, data []interface{}) ([]string, error) {
+func (m *RepositoryDummy) InsertMany(
+	ctx context.Context,
+	collection mongoModel.Collection,
+	data []any,
+) ([]string, error) {
 	return nil, nil
 }
 
-func (m *RepositoryDummy) FindOneAndUpdate(ctx context.Context, collection mongoModel.Collection, resultModel, filter, updateData interface{}, opt *options.FindOneAndUpdateOptions) error {
+func (m *RepositoryDummy) FindOneAndUpdate(
+	ctx context.Context,
+	collection mongoModel.Collection,
+	resultModel, filter,
+	updateData any,
+	opt *options.FindOneAndUpdateOptions,
+) error {
 	return nil
 }
 
-func (m *RepositoryDummy) ReplaceOne(ctx context.Context, collection mongoModel.Collection, filter, data interface{}) error {
+func (m *RepositoryDummy) ReplaceOne(
+	ctx context.Context,
+	collection mongoModel.Collection,
+	filter, data any,
+) error {
 	return nil
 }
 
-func (m *RepositoryDummy) UpdateOne(ctx context.Context, collection mongoModel.Collection, filter, data interface{}, opts ...*options.UpdateOptions) (int64, error) {
+func (m *RepositoryDummy) UpdateOne(
+	ctx context.Context,
+	collection mongoModel.Collection,
+	filter, data any,
+	opts ...options.Lister[options.UpdateOneOptions],
+) (int64, error) {
 	return 0, nil
 }
 
 func (m *RepositoryDummy) UpdateMany(
 	ctx context.Context,
 	collection mongoModel.Collection,
-	filter interface{},
-	data interface{},
-	opts ...*options.UpdateOptions,
+	filter any,
+	data any,
+	opts ...options.Lister[options.UpdateManyOptions],
 ) (int64, error) {
 	return 0, nil
 }
 
-func (m *RepositoryDummy) Find(ctx context.Context, collection mongoModel.Collection, results, find interface{}, opt *options.FindOptions) error {
+func (m *RepositoryDummy) Find(
+	ctx context.Context,
+	collection mongoModel.Collection,
+	results, find any,
+	opt *options.FindOptions,
+) error {
 	return nil
 }
 
-func (m *RepositoryDummy) FindOne(ctx context.Context, collection mongoModel.Collection, resultModel, findQuery interface{}, findOptions *options.FindOneOptions) error {
+func (m *RepositoryDummy) FindOne(
+	ctx context.Context,
+	collection mongoModel.Collection,
+	resultModel, findQuery any,
+	findOptions *options.FindOneOptions,
+) error {
 	return nil
 }
 
-func (m *RepositoryDummy) DeleteOne(ctx context.Context, collection mongoModel.Collection, filter interface{}, opt *options.DeleteOptions) (*mongo.DeleteResult, error) {
+func (m *RepositoryDummy) DeleteOne(
+	ctx context.Context,
+	collection mongoModel.Collection,
+	filter any,
+	opts ...options.Lister[options.DeleteOneOptions],
+) (*mongo.DeleteResult, error) {
 	return nil, nil
 }
 
-func (m *RepositoryDummy) DeleteMany(ctx context.Context, collection mongoModel.Collection, filter interface{}, opt *options.DeleteOptions) (*mongo.DeleteResult, error) {
+func (m *RepositoryDummy) DeleteMany(
+	ctx context.Context,
+	collection mongoModel.Collection,
+	filter any,
+	opts ...options.Lister[options.DeleteManyOptions],
+) (*mongo.DeleteResult, error) {
 	return nil, nil
 }
 
-func (m *RepositoryDummy) Count(ctx context.Context, collection mongoModel.Collection, find interface{}, opt *options.CountOptions) (int64, error) {
+func (m *RepositoryDummy) Count(
+	ctx context.Context,
+	collection mongoModel.Collection,
+	find any,
+	opt *options.CountOptions,
+) (int64, error) {
 	return 0, nil
 }
 
-func (m *RepositoryDummy) Aggregate(ctx context.Context, collection mongoModel.Collection, pipe mongo.Pipeline) (*mongo.Cursor, error) {
+func (m *RepositoryDummy) Aggregate(
+	ctx context.Context,
+	collection mongoModel.Collection,
+	pipe mongo.Pipeline,
+) (*mongo.Cursor, error) {
 	return nil, nil
 }
 
-func (m *RepositoryDummy) CreateIndex(ctx context.Context, index *mongoModel.DBIndex) (string, error) {
+func (m *RepositoryDummy) CreateIndex(
+	ctx context.Context,
+	index *mongoModel.DBIndex,
+) (string, error) {
 	return "", nil
 }
 
-func (m *RepositoryDummy) CreateTextIndex(ctx context.Context, index *mongoModel.DBTextIndex) (string, error) {
+func (m *RepositoryDummy) CreateTextIndex(
+	ctx context.Context,
+	index *mongoModel.DBTextIndex,
+) (string, error) {
 	return "", nil
 }
 
-func (m *RepositoryDummy) CollectionIndexes(ctx context.Context, collection mongoModel.Collection) (map[string]*mongoModel.DBIndex, error) {
+func (m *RepositoryDummy) CollectionIndexes(
+	ctx context.Context,
+	collection mongoModel.Collection,
+) (map[string]*mongoModel.DBIndex, error) {
 	return nil, nil
 }
 
-func (m *RepositoryDummy) TryCreateIndex(ctx context.Context, index *mongoModel.DBIndex) error {
+func (m *RepositoryDummy) TryCreateIndex(
+	ctx context.Context,
+	index *mongoModel.DBIndex,
+) error {
 	return nil
 }
