@@ -2,11 +2,12 @@ package mongoMock
 
 import (
 	"context"
+
 	mocking "github.com/conacry/go-platform/pkg/mock"
 	mongoModel "github.com/conacry/go-platform/pkg/mongo/model"
 
-	"go.mongodb.org/mongo-driver/mongo"
-	"go.mongodb.org/mongo-driver/mongo/options"
+	"go.mongodb.org/mongo-driver/v2/mongo"
+	"go.mongodb.org/mongo-driver/v2/mongo/options"
 )
 
 func GetRepository() *RepositoryMock {
@@ -21,7 +22,11 @@ type RepositoryMock struct {
 	*mocking.BaseMock
 }
 
-func (m *RepositoryMock) Insert(ctx context.Context, collection mongoModel.Collection, data interface{}) (string, error) {
+func (m *RepositoryMock) Insert(
+	ctx context.Context,
+	collection mongoModel.Collection,
+	data any,
+) (string, error) {
 	result, err := m.ProcessMethod(ctx, collection, data)
 	if err != nil {
 		return "", err
@@ -39,7 +44,11 @@ func (m *RepositoryMock) Insert(ctx context.Context, collection mongoModel.Colle
 	return "", err
 }
 
-func (m *RepositoryMock) InsertMany(ctx context.Context, collection mongoModel.Collection, data []interface{}) ([]string, error) {
+func (m *RepositoryMock) InsertMany(
+	ctx context.Context,
+	collection mongoModel.Collection,
+	data []any,
+) ([]string, error) {
 	result, err := m.ProcessMethod(ctx, collection, data)
 	if err != nil {
 		return nil, err
@@ -57,17 +66,32 @@ func (m *RepositoryMock) InsertMany(ctx context.Context, collection mongoModel.C
 	return nil, err
 }
 
-func (m *RepositoryMock) FindOneAndUpdate(ctx context.Context, collection mongoModel.Collection, resultModel, filter, updateData interface{}, opt *options.FindOneAndUpdateOptions) error {
+func (m *RepositoryMock) FindOneAndUpdate(
+	ctx context.Context,
+	collection mongoModel.Collection,
+	resultModel, filter,
+	updateData any,
+	opt *options.FindOneAndUpdateOptions,
+) error {
 	_, err := m.ProcessMethod(ctx, collection, resultModel, filter, updateData, opt)
 	return err
 }
 
-func (m *RepositoryMock) ReplaceOne(ctx context.Context, collection mongoModel.Collection, filter, data interface{}) error {
+func (m *RepositoryMock) ReplaceOne(
+	ctx context.Context,
+	collection mongoModel.Collection,
+	filter, data any,
+) error {
 	_, err := m.ProcessMethod(ctx, collection, filter, data)
 	return err
 }
 
-func (m *RepositoryMock) UpdateOne(ctx context.Context, collection mongoModel.Collection, filter, data interface{}, opts ...*options.UpdateOptions) (int64, error) {
+func (m *RepositoryMock) UpdateOne(
+	ctx context.Context,
+	collection mongoModel.Collection,
+	filter, data any,
+	opts ...options.Lister[options.UpdateOneOptions],
+) (int64, error) {
 	result, err := m.ProcessMethod(ctx, collection, filter, data, opts)
 	if err != nil {
 		return 0, err
@@ -88,9 +112,8 @@ func (m *RepositoryMock) UpdateOne(ctx context.Context, collection mongoModel.Co
 func (m *RepositoryMock) UpdateMany(
 	ctx context.Context,
 	collection mongoModel.Collection,
-	filter interface{},
-	data interface{},
-	opts ...*options.UpdateOptions,
+	filter, data any,
+	opts ...options.Lister[options.UpdateManyOptions],
 ) (int64, error) {
 	result, err := m.ProcessMethod(ctx, collection, filter, data, opts)
 	if err != nil {
@@ -109,18 +132,37 @@ func (m *RepositoryMock) UpdateMany(
 	return 0, err
 }
 
-func (m *RepositoryMock) Find(ctx context.Context, collection mongoModel.Collection, results, find interface{}, opt *options.FindOptions) error {
+func (m *RepositoryMock) Find(
+	ctx context.Context,
+	collection mongoModel.Collection,
+	results, find any,
+	opt *options.FindOptions,
+) error {
 	_, err := m.ProcessMethod(ctx, collection, results, find, opt)
 	return err
 }
 
-func (m *RepositoryMock) FindOne(ctx context.Context, collection mongoModel.Collection, resultModel, findQuery interface{}, findOptions *options.FindOneOptions) error {
-	_, err := m.ProcessMethod(ctx, collection, resultModel, findQuery, findOptions)
+func (m *RepositoryMock) FindOne(
+	ctx context.Context,
+	collection mongoModel.Collection,
+	resultModel, findQuery any,
+	opts ...options.Lister[options.FindOneOptions],
+) error {
+	args := []any{ctx, collection, resultModel, findQuery}
+	args = append(args, opts)
+	_, err := m.ProcessMethod(args...)
 	return err
 }
 
-func (m *RepositoryMock) DeleteOne(ctx context.Context, collection mongoModel.Collection, filter interface{}, opt *options.DeleteOptions) (*mongo.DeleteResult, error) {
-	result, err := m.ProcessMethod(ctx, collection, filter, opt)
+func (m *RepositoryMock) DeleteOne(
+	ctx context.Context,
+	collection mongoModel.Collection,
+	filter any,
+	opts ...options.Lister[options.DeleteOneOptions],
+) (*mongo.DeleteResult, error) {
+	args := []any{ctx, collection, filter}
+	args = append(args, opts)
+	result, err := m.ProcessMethod(args...)
 	if err != nil {
 		return nil, err
 	}
@@ -137,8 +179,15 @@ func (m *RepositoryMock) DeleteOne(ctx context.Context, collection mongoModel.Co
 	return nil, err
 }
 
-func (m *RepositoryMock) DeleteMany(ctx context.Context, collection mongoModel.Collection, filter interface{}, opt *options.DeleteOptions) (*mongo.DeleteResult, error) {
-	result, err := m.ProcessMethod(ctx, collection, filter, opt)
+func (m *RepositoryMock) DeleteMany(
+	ctx context.Context,
+	collection mongoModel.Collection,
+	filter any,
+	opts ...options.Lister[options.DeleteManyOptions],
+) (*mongo.DeleteResult, error) {
+	args := []any{ctx, collection, filter}
+	args = append(args, opts)
+	result, err := m.ProcessMethod(args...)
 	if err != nil {
 		return nil, err
 	}
@@ -155,7 +204,12 @@ func (m *RepositoryMock) DeleteMany(ctx context.Context, collection mongoModel.C
 	return nil, err
 }
 
-func (m *RepositoryMock) Count(ctx context.Context, collection mongoModel.Collection, find interface{}, opt *options.CountOptions) (int64, error) {
+func (m *RepositoryMock) Count(
+	ctx context.Context,
+	collection mongoModel.Collection,
+	find any,
+	opt *options.CountOptions,
+) (int64, error) {
 	result, err := m.ProcessMethod(ctx, collection, find, opt)
 	if err != nil {
 		return 0, err
@@ -173,7 +227,11 @@ func (m *RepositoryMock) Count(ctx context.Context, collection mongoModel.Collec
 	return 0, err
 }
 
-func (m *RepositoryMock) Aggregate(ctx context.Context, collection mongoModel.Collection, pipe mongo.Pipeline) (*mongo.Cursor, error) {
+func (m *RepositoryMock) Aggregate(
+	ctx context.Context,
+	collection mongoModel.Collection,
+	pipe mongo.Pipeline,
+) (*mongo.Cursor, error) {
 	result, err := m.ProcessMethod(ctx, collection, pipe)
 	if err != nil {
 		return nil, err
@@ -232,7 +290,10 @@ func (m *RepositoryMock) TryCreateIndex(ctx context.Context, index *mongoModel.D
 	return err
 }
 
-func (m *RepositoryMock) CollectionIndexes(ctx context.Context, collection mongoModel.Collection) (map[string]*mongoModel.DBIndex, error) {
+func (m *RepositoryMock) CollectionIndexes(
+	ctx context.Context,
+	collection mongoModel.Collection,
+) (map[string]*mongoModel.DBIndex, error) {
 	result, err := m.ProcessMethod(ctx, collection)
 	if err != nil {
 		return nil, err
