@@ -2,8 +2,8 @@ package integration
 
 import (
 	"github.com/stretchr/testify/assert"
-	"go.mongodb.org/mongo-driver/bson"
-	"go.mongodb.org/mongo-driver/mongo/options"
+	"go.mongodb.org/mongo-driver/v2/bson"
+	"go.mongodb.org/mongo-driver/v2/mongo/options"
 )
 
 func (s *MongoContainerShould) TestFind_WithNoFilterAndAscendingOrder_ReturnSuccess() {
